Add helpers.js file to handle long .then() statement
